refactor: reorganize storage-related objects
