+++ errors/src/objects/storage.rs
@@ -14,13 +14,11 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use snarkvm_errors::{
+use crate::{
     algorithms::MerkleError,
     objects::{BlockError, TransactionError},
     parameters::ParameterError,
 };
-
-use std::fmt::Debug;
 
 #[derive(Debug, Error)]
 pub enum StorageError {
@@ -135,19 +133,7 @@ pub enum StorageError {
 
 impl From<bincode::Error> for StorageError {
     fn from(error: bincode::Error) -> Self {
-        StorageError::Crate("bincode", format!("{:?}", error))
-    }
-}
-
-impl From<hex::FromHexError> for StorageError {
-    fn from(error: hex::FromHexError) -> Self {
-        StorageError::Crate("hex", format!("{:?}", error))
-    }
-}
-
-impl From<rocksdb::Error> for StorageError {
-    fn from(error: rocksdb::Error) -> Self {
-        StorageError::Crate("rocksdb", format!("{:?}", error))
+        StorageError::Crate("bincode", error.to_string())
     }
 }
