+++ models/src/objects/storage.rs
@@ -1,0 +1,76 @@
+// Copyright (C) 2019-2021 Aleo Systems Inc.
+// This file is part of the snarkVM library.
+
+// The snarkVM library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// The snarkVM library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
+
+use snarkvm_errors::objects::StorageError;
+
+use std::{fmt::Debug, path::Path};
+
+pub trait Storage
+where
+    Self: Sized,
+{
+    /// Opens the storage object, optionally using the given paths; it gets created if it doesn't exist.
+    fn open(path: Option<&Path>, secondary_path: Option<&Path>) -> Result<Self, StorageError>;
+
+    /// Returns the value with the given key and belonging to the given column.
+    fn get(&self, col: u32, key: &[u8]) -> Result<Option<Vec<u8>>, StorageError>;
+
+    /// Returns all the keys and values belonging to the given column.
+    fn get_col(&self, col: u32) -> Result<Vec<(Box<[u8]>, Box<[u8]>)>, StorageError>;
+
+    /// Returns all the keys belonging to the given column.
+    fn get_keys(&self, col: u32) -> Result<Vec<Box<[u8]>>, StorageError>;
+
+    /// Stores the given key and value in the specified column.
+    fn put<K: AsRef<[u8]>, V: AsRef<[u8]>>(&self, col: u32, key: K, value: V) -> Result<(), StorageError>;
+
+    /// Executes the given `StorageBatchOp` as a batch operation.
+    fn batch(&self, batch: StorageBatchOp) -> Result<(), StorageError>;
+
+    /// Returns `true` if the given key exists in the speficied column.
+    fn exists(&self, col: u32, key: &[u8]) -> bool;
+
+    /// Used to remove any persistent objects associated with the storage.
+    fn destroy(&self) -> Result<(), StorageError>;
+}
+
+/// Database operation.
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub enum StorageOp {
+    Insert { col: u32, key: Vec<u8>, value: Vec<u8> },
+    Delete { col: u32, key: Vec<u8> },
+}
+
+/// Batched transaction of database operations.
+#[derive(Default, Clone, PartialEq)]
+pub struct StorageBatchOp(pub Vec<StorageOp>);
+
+impl StorageBatchOp {
+    /// Create new transaction.
+    pub fn new() -> Self {
+        Self(vec![])
+    }
+
+    /// Add an operation.
+    pub fn push(&mut self, op: StorageOp) {
+        self.0.push(op)
+    }
+
+    /// Add a vector of operations.
+    pub fn push_vec(&mut self, ops: Vec<StorageOp>) {
+        self.0.extend(ops)
+    }
+}
