+++ storage/Cargo.toml
@@ -56,9 +56,6 @@ version = "0.11.1"
 [dependencies.rand]
 version = "0.8"
 
-[dependencies.rocksdb]
-version = "0.15.0"
-
 [dependencies.serde]
 version = "1.0"
 features = [ "derive" ]
