+++ storage/src/key_value.rs
@@ -66,73 +66,3 @@ impl FromBytes for TransactionLocation {
         Ok(Self { index, block_hash })
     }
 }
-
-/// Database operation.
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub enum Op {
-    Insert { col: u32, key: Vec<u8>, value: Vec<u8> },
-    Delete { col: u32, key: Vec<u8> },
-}
-
-impl Op {
-    pub fn key(&self) -> &[u8] {
-        match self {
-            Op::Insert { key, .. } => &key,
-            Op::Delete { key, .. } => &key,
-        }
-    }
-
-    pub fn col(&self) -> u32 {
-        match self {
-            Op::Insert { col, .. } => *col,
-            Op::Delete { col, .. } => *col,
-        }
-    }
-}
-
-/// Batched transaction of database operations.
-#[derive(Default, Clone, PartialEq)]
-pub struct DatabaseTransaction(pub Vec<Op>);
-
-impl DatabaseTransaction {
-    /// Create new transaction.
-    pub fn new() -> Self {
-        Self(vec![])
-    }
-
-    /// Add a key value pair under a specific col.
-    pub fn add(&mut self, col: u32, key: &[u8], value: &[u8]) {
-        self.0.push(Op::Insert {
-            col,
-            key: key.to_vec(),
-            value: value.to_vec(),
-        })
-    }
-
-    /// Delete a value given a col and key.
-    pub fn delete(&mut self, col: u32, key: &[u8]) {
-        self.0.push(Op::Delete { col, key: key.to_vec() })
-    }
-
-    /// Add an operation.
-    pub fn push(&mut self, op: Op) {
-        self.0.push(op)
-    }
-
-    /// Add a vector of operations.
-    pub fn push_vec(&mut self, ops: Vec<Op>) {
-        self.0.extend(ops)
-    }
-
-    /// Add another database transaction.
-    pub fn extend(&mut self, database_transaction: DatabaseTransaction) {
-        self.0.extend(database_transaction.0)
-    }
-}
-
-pub fn bytes_to_u32(bytes: Vec<u8>) -> u32 {
-    let mut num_bytes = [0u8; 4];
-    num_bytes.copy_from_slice(&bytes);
-
-    u32::from_le_bytes(num_bytes)
-}
