+++ storage/src/lib.rs
@@ -14,11 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-#[macro_use]
-extern crate thiserror;
-
-pub mod error;
-
 pub mod ledger;
 pub use ledger::*;
 
@@ -27,6 +22,3 @@ pub use key_value::*;
 
 pub mod objects;
 pub use objects::*;
-
-pub mod storage;
-pub use storage::*;
