+++ testing/Cargo.toml
@@ -56,6 +56,9 @@ version = "0.0.4"
 [dependencies.once_cell]
 version = "1.5.2"
 
+[dependencies.parking_lot]
+version = "0.11"
+
 [dependencies.rand]
 version = "0.8"
