+++ testing/src/storage/rocks.rs
@@ -14,101 +14,15 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{error::StorageError, DatabaseTransaction, Op, NUM_COLS};
+use snarkvm_errors::objects::StorageError;
+use snarkvm_models::objects::{Storage, StorageBatchOp, StorageOp};
+use snarkvm_storage::NUM_COLS;
 
-use parking_lot::RwLock;
 use rocksdb::{ColumnFamily, ColumnFamilyDescriptor, IteratorMode, Options, WriteBatch, DB};
-use std::{collections::HashMap, path::Path};
+use std::path::Path;
 
-pub trait Storage
-where
-    Self: Sized,
-{
-    /// Opens the storage object, optionally using the given paths; it gets created if it doesn't exist.
-    fn open(path: Option<&Path>, secondary_path: Option<&Path>) -> Result<Self, StorageError>;
-
-    /// Returns the value with the given key and belonging to the given column.
-    fn get(&self, col: u32, key: &[u8]) -> Result<Option<Vec<u8>>, StorageError>;
-
-    /// Returns all the keys and values belonging to the given column.
-    fn get_col(&self, col: u32) -> Result<Vec<(Box<[u8]>, Box<[u8]>)>, StorageError>;
-
-    /// Returns all the keys belonging to the given column.
-    fn get_keys(&self, col: u32) -> Result<Vec<Box<[u8]>>, StorageError>;
-
-    /// Stores the given key and value in the specified column.
-    fn put<K: AsRef<[u8]>, V: AsRef<[u8]>>(&self, col: u32, key: K, value: V) -> Result<(), StorageError>;
-
-    /// Executes the given `DatabaseTransaction` in a single batch operation.
-    fn write(&self, transaction: DatabaseTransaction) -> Result<(), StorageError>;
-
-    /// Returns `true` if the given key exists in the speficied column.
-    fn exists(&self, col: u32, key: &[u8]) -> bool;
-
-    /// Used to remove any persistent objects associated with the storage.
-    fn destroy(&self) -> Result<(), StorageError>;
-}
-
-pub struct MemDb {
-    pub cols: RwLock<Vec<HashMap<Box<[u8]>, Box<[u8]>>>>,
-}
-
-impl Storage for MemDb {
-    fn open(path: Option<&Path>, secondary_path: Option<&Path>) -> Result<Self, StorageError> {
-        assert!(
-            path.is_none() && secondary_path.is_none(),
-            "MemDb has no associated filesystem paths!"
-        );
-
-        Ok(Self {
-            cols: RwLock::new(vec![Default::default(); NUM_COLS as usize]),
-        })
-    }
-
-    fn get(&self, col: u32, key: &[u8]) -> Result<Option<Vec<u8>>, StorageError> {
-        Ok(self.cols.read()[col as usize].get(key).map(|v| v.to_vec()))
-    }
-
-    fn get_col(&self, col: u32) -> Result<Vec<(Box<[u8]>, Box<[u8]>)>, StorageError> {
-        Ok(self.cols.read()[col as usize].clone().into_iter().collect())
-    }
-
-    fn get_keys(&self, col: u32) -> Result<Vec<Box<[u8]>>, StorageError> {
-        Ok(self.cols.read()[col as usize].keys().cloned().collect())
-    }
-
-    fn put<K: AsRef<[u8]>, V: AsRef<[u8]>>(&self, col: u32, key: K, value: V) -> Result<(), StorageError> {
-        self.cols.write()[col as usize].insert(key.as_ref().into(), value.as_ref().into());
-        Ok(())
-    }
-
-    fn write(&self, transaction: DatabaseTransaction) -> Result<(), StorageError> {
-        if transaction.0.is_empty() {
-            return Ok(());
-        }
-
-        let mut cols = self.cols.write();
-        for operation in transaction.0 {
-            match operation {
-                Op::Insert { col, key, value } => {
-                    cols[col as usize].insert(key.into(), value.into());
-                }
-                Op::Delete { col, key } => {
-                    cols[col as usize].remove(&Box::from(key));
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    fn exists(&self, col: u32, key: &[u8]) -> bool {
-        self.cols.read()[col as usize].contains_key(key)
-    }
-
-    fn destroy(&self) -> Result<(), StorageError> {
-        Ok(()) // nothing to do here: the Drop impl takes care of the associated in-memory objects
-    }
+fn convert_err(err: rocksdb::Error) -> StorageError {
+    StorageError::Crate("rocksdb", err.to_string())
 }
 
 /// A low-level struct for storing state used by the system.
@@ -130,7 +44,7 @@ impl Storage for RocksDb {
     }
 
     fn get(&self, col: u32, key: &[u8]) -> Result<Option<Vec<u8>>, StorageError> {
-        Ok(self.db.get_cf(self.get_cf_ref(col), key)?)
+        Ok(self.db.get_cf(self.get_cf_ref(col), key).map_err(convert_err)?)
     }
 
     fn get_col(&self, col: u32) -> Result<Vec<(Box<[u8]>, Box<[u8]>)>, StorageError> {
@@ -146,26 +60,26 @@ impl Storage for RocksDb {
     }
 
     fn put<K: AsRef<[u8]>, V: AsRef<[u8]>>(&self, col: u32, key: K, value: V) -> Result<(), StorageError> {
-        Ok(self.db.put_cf(self.get_cf_ref(col), key, value)?)
+        Ok(self.db.put_cf(self.get_cf_ref(col), key, value).map_err(convert_err)?)
     }
 
-    fn write(&self, transaction: DatabaseTransaction) -> Result<(), StorageError> {
+    fn batch(&self, transaction: StorageBatchOp) -> Result<(), StorageError> {
         let mut batch = WriteBatch::default();
 
         for operation in transaction.0 {
             match operation {
-                Op::Insert { col, key, value } => {
+                StorageOp::Insert { col, key, value } => {
                     let cf = self.get_cf_ref(col);
                     batch.put_cf(cf, &key, value);
                 }
-                Op::Delete { col, key } => {
+                StorageOp::Delete { col, key } => {
                     let cf = self.get_cf_ref(col);
                     batch.delete_cf(cf, &key);
                 }
             };
         }
 
-        self.db.write(batch)?;
+        self.db.write(batch).map_err(convert_err)?;
 
         Ok(())
     }
@@ -184,7 +98,7 @@ impl Storage for RocksDb {
         storage_opts.create_missing_column_families(true);
         storage_opts.create_if_missing(true);
 
-        Ok(DB::destroy(&storage_opts, path)?)
+        Ok(DB::destroy(&storage_opts, path).map_err(convert_err)?)
     }
 }
 
@@ -217,7 +131,7 @@ impl RocksDb {
         storage_opts.create_missing_column_families(true);
         storage_opts.create_if_missing(true);
 
-        let storage = DB::open_cf_descriptors(&storage_opts, path, cfs)?;
+        let storage = DB::open_cf_descriptors(&storage_opts, path, cfs).map_err(convert_err)?;
 
         Ok(Self { db: storage, cf_names })
     }
@@ -240,9 +154,10 @@ impl RocksDb {
         let mut storage_opts = Options::default();
         storage_opts.increase_parallelism(2);
 
-        let storage = DB::open_cf_as_secondary(&storage_opts, primary_path, secondary_path, cf_names.clone())?;
+        let storage = DB::open_cf_as_secondary(&storage_opts, primary_path, secondary_path, cf_names.clone())
+            .map_err(convert_err)?;
 
-        storage.try_catch_up_with_primary()?;
+        storage.try_catch_up_with_primary().map_err(convert_err)?;
 
         Ok(Self { db: storage, cf_names })
     }
